+++ src/Building.h
@@ -52,7 +52,6 @@ public:
     void   clip_bottom(const TerrainPtr& terrain);
     void   refine();
     void   alpha_wrap(double relativeAlpha, double relativeOffset);
-    void   translate_footprint(const double h);
     bool   is_part_of(const Polygon_2& otherPoly) const;
     void   set_reconstruction_rules(const BoundingRegion& reconRegion);
     void   remove_reconstruction_rules();
@@ -80,6 +79,8 @@ protected:
     bool                 m_hasFailed;
     bool                 m_clipBottom = Config::get().clip;
     std::shared_ptr<const Config::ReconRegion> m_reconSettings;
+
+    void force_building_terrain_intersection(const double h = 10.);
 };
 
 //-- Struct for clipping