+++ src/ImportedBuilding.cpp
@@ -291,9 +291,6 @@ void ImportedBuilding::reconstruct() {
     nlohmann::json& geometry = (*m_buildingJson)["geometry"][m_lodIdx];
 
     m_mesh.clear();
-    if (m_clipBottom || Config::get().intersectBuildingsTerrain) {
-        this->translate_footprint(-5);
-    }
     //-- Adjust building height points
     if (!m_trueHeight || Config::get().ground_xyz.empty()) {
         Vector_3 movePt(0, 0, this->ground_elevation());
@@ -370,9 +367,6 @@ void ImportedBuilding::reconstruct() {
     m_mesh = wrap;
      */
 
-    if (m_clipBottom || Config::get().intersectBuildingsTerrain) {
-        this->translate_footprint(5);
-    }
         /*
         //-- Add other surfaces
         std::vector<Mesh::Vertex_index> faceVertices;
@@ -401,6 +395,9 @@ void ImportedBuilding::reconstruct() {
     PMP::triangulate_faces(m_mesh);
     */
     if (Config::get().refineImported) this->refine();
+    if (m_clipBottom || Config::get().intersectBuildingsTerrain) {
+        this->force_building_terrain_intersection(5);
+    }
 }
 
 void ImportedBuilding::reconstruct_flat_terrain() {